crud feito
tentando fazer a query do express

+++ WebAPI_atv13/src/microblog.ts
@@ -1,14 +1,11 @@
-import { Any } from "typeorm"
 import { PostFire } from "./firestore.config"
 
-class Post {
-  id: string
+export class Post {
   text: string
   likes: number
   date: Date
 
   constructor(id: string, text: string, likes: number, date: Date) {
-      this.id = id
       this.text = text
       this.likes = likes
       this.date = date
@@ -17,10 +14,12 @@ class Post {
 
 export class microblog{
 
+  post: Post[] = []
+
   // Criar um novo post
   async create (texto: string){
     const criar = await PostFire.add({
-      texto,
+      text: texto,
       likes : 0,
       date : new Date()
     })
@@ -29,33 +28,55 @@ export class microblog{
   }
   
   // Retornar todos os posts ordenados pela data mais recente para mais antiga
-  async getAll (){
-    const snapshot = await PostFire.get()
+  async getAll (query: any): Promise<Post[]>{
+    const snapshot = await PostFire.where('date','<=',new Date()).get()
+    
+    let altura: number = this.post.length
+    this.post.splice(0,altura)
 
-    snapshot.forEach(function (doc: any){
-      console.log(doc.id,'=>',doc.data())
+    const vazio: any = {}
+    console.log(query)
+
+    snapshot.forEach((doc: any) => {
+      this.post.push(doc.data())
     })
+      
+    return this.post
   }
 
   // Retornar apenas um post pelo id recebido
   async getById(id: string){
     const post = await PostFire.doc(id).get()
-    const postReturn = post.data()
+    
+    let altura: number = this.post.length
+    this.post.splice(0,altura)
 
-    return postReturn
+    this.post.push(post.data())
+
+    return this.post
   }
 
   // Fazer update dos posts através do id
-  async update (id: string,alt: string){
-    const idRef = PostFire.doc(id)
-    
-    await idRef.update({
-      text : alt
-    })
+  async update (id: string,alt: string): Promise<boolean>{
+    const postRef = await PostFire.doc(id).get()
+
+    if(postRef.data() != null){
+      await PostFire.doc(id).update({text : alt})
+      return true
+    }else{
+      return false
+    }
   }
   
   // Deletar através do id
-  async delete (id: string){
-    const idRef = PostFire.doc(id).delete()
+  async delete (id: string): Promise<boolean>{
+    const idRef = await PostFire.doc(id).get()
+
+    if(idRef.data() != null){
+      PostFire.doc(id).delete()
+      return true
+    }else{
+      return false
+    }
   }
 }
